Exposed some methods of config.
